+++ packages/core/tests/crameri-ramps-continuous-data.ts
@@ -1,0 +1,293 @@
+/**
+ * Crameri sequential (continuous) ramp fixtures — 11-stop samples
+ * from Scientific Colour Maps v8.0.1 (Zenodo DOI 10.5281/zenodo.1243862).
+ * Data only: never imports bun:test.
+ */
+export const CONTINUOUS_RAMPS: Readonly<Record<string, readonly string[]>> = {
+  acton: [
+    "#260d40",
+    "#3c2c5c",
+    "#4f4775",
+    "#665e8a",
+    "#85648e",
+    "#a86690",
+    "#ca7199",
+    "#d890b3",
+    "#e2b2cf",
+    "#e9d1e7",
+    "#f0eafa",
+  ],
+  bamako: [
+    "#003b47",
+    "#0d4340",
+    "#1b4c37",
+    "#2e582b",
+    "#47681c",
+    "#637a0a",
+    "#818800",
+    "#a2930d",
+    "#c6ae39",
+    "#e5cb75",
+    "#ffe5ad",
+  ],
+  batlow: [
+    "#011959",
+    "#103d5f",
+    "#185562",
+    "#30685c",
+    "#577647",
+    "#828231",
+    "#b38e2f",
+    "#e09651",
+    "#fba689",
+    "#fdb9c2",
+    "#faccfa",
+  ],
+  batlowW: [
+    "#011959",
+    "#103d5f",
+    "#175462",
+    "#2d685e",
+    "#53784c",
+    "#7e8737",
+    "#b1993a",
+    "#daa66b",
+    "#f3b49e",
+    "#ffdddb",
+    "#fffefe",
+  ],
+  batlowK: [
+    "#04050a",
+    "#18293d",
+    "#304d5d",
+    "#47605c",
+    "#61704d",
+    "#86833c",
+    "#b7953c",
+    "#e4a15f",
+    "#f9ac92",
+    "#fdbbc5",
+    "#faccfa",
+  ],
+  bilbao: [
+    "#4c0001",
+    "#712227",
+    "#8e3f46",
+    "#9e5a55",
+    "#a46f5a",
+    "#a9825e",
+    "#ae9663",
+    "#b9ae80",
+    "#c4c0ac",
+    "#d6d5d3",
+    "#ffffff",
+  ],
+  buda: [
+    "#b301b3",
+    "#b3289f",
+    "#b84294",
+    "#bf598c",
+    "#c76f85",
+    "#cd857e",
+    "#d39b79",
+    "#d8b173",
+    "#dec96d",
+    "#e7e268",
+    "#ffff66",
+  ],
+  davos: [
+    "#00054a",
+    "#102a6f",
+    "#234a8c",
+    "#3a679b",
+    "#547d9c",
+    "#6c8e93",
+    "#849e89",
+    "#a4b68a",
+    "#d4dba8",
+    "#f5f5d7",
+    "#fefefe",
+  ],
+  devon: [
+    "#2c1a4c",
+    "#293265",
+    "#274a7e",
+    "#2f62a0",
+    "#4c77c4",
+    "#7e8fdd",
+    "#a8a5ec",
+    "#c0baf2",
+    "#d5d1f6",
+    "#eae8fb",
+    "#ffffff",
+  ],
+  glasgow: [
+    "#361338",
+    "#4d1924",
+    "#652211",
+    "#743b01",
+    "#735704",
+    "#6d702d",
+    "#65845d",
+    "#60988c",
+    "#7eaeb9",
+    "#abc0dc",
+    "#dbd3ff",
+  ],
+  grayC: [
+    "#000000",
+    "#212121",
+    "#383838",
+    "#4e4e4e",
+    "#636363",
+    "#777777",
+    "#8b8b8b",
+    "#a2a2a2",
+    "#bdbdbd",
+    "#dcdcdc",
+    "#ffffff",
+  ],
+  hawaii: [
+    "#8c0273",
+    "#91285a",
+    "#954147",
+    "#985936",
+    "#9c7524",
+    "#9c961c",
+    "#8fb63c",
+    "#78cb70",
+    "#62dca9",
+    "#71ece0",
+    "#b3f2fd",
+  ],
+  imola: [
+    "#1a33b3",
+    "#2345aa",
+    "#2c55a1",
+    "#356598",
+    "#42748b",
+    "#54867f",
+    "#6a9d78",
+    "#83b672",
+    "#9fd26b",
+    "#caec67",
+    "#ffff66",
+  ],
+  lajolla: [
+    "#191900",
+    "#31220e",
+    "#512d1e",
+    "#7d3b35",
+    "#b34947",
+    "#d9604e",
+    "#e38050",
+    "#e99d53",
+    "#f0bd57",
+    "#f9e384",
+    "#fffecb",
+  ],
+  lapaz: [
+    "#1a0c64",
+    "#222b79",
+    "#28468b",
+    "#315e99",
+    "#4277a2",
+    "#5c8ca3",
+    "#7c9b9e",
+    "#9fa696",
+    "#cab79d",
+    "#f2d7c6",
+    "#fef2f3",
+  ],
+  lipari: [
+    "#031326",
+    "#103557",
+    "#3c5478",
+    "#615e78",
+    "#806070",
+    "#a56267",
+    "#cf695e",
+    "#e98768",
+    "#e6aa7f",
+    "#eacda5",
+    "#fdf5da",
+  ],
+  navia: [
+    "#031327",
+    "#06315a",
+    "#105185",
+    "#236c91",
+    "#327c89",
+    "#418a80",
+    "#549b75",
+    "#70b369",
+    "#a7d278",
+    "#dde7ae",
+    "#fcf4d9",
+  ],
+  naviaW: [
+    "#041427",
+    "#08345c",
+    "#165584",
+    "#29708f",
+    "#388188",
+    "#4a927f",
+    "#62a774",
+    "#8cc574",
+    "#cbe6a3",
+    "#eff6db",
+    "#fefefd",
+  ],
+  nuuk: [
+    "#05598c",
+    "#276184",
+    "#436e82",
+    "#638089",
+    "#859493",
+    "#a1a698",
+    "#b2b293",
+    "#bdbd8a",
+    "#caca83",
+    "#e4e391",
+    "#fefeb2",
+  ],
+  oslo: [
+    "#010101",
+    "#0d1927",
+    "#122d48",
+    "#1a446d",
+    "#2c5d96",
+    "#507bbc",
+    "#7494c9",
+    "#92a6c9",
+    "#b2bccc",
+    "#d8dade",
+    "#ffffff",
+  ],
+  tokyo: [
+    "#1c0e34",
+    "#481e43",
+    "#653a4d",
+    "#6e5151",
+    "#716152",
+    "#747053",
+    "#788555",
+    "#7fa35c",
+    "#95cb78",
+    "#c7f0b2",
+    "#effcdd",
+  ],
+  turku: [
+    "#000000",
+    "#22221f",
+    "#3c3c32",
+    "#565640",
+    "#71704c",
+    "#938c5b",
+    "#b9a06e",
+    "#d8a782",
+    "#efb1a1",
+    "#fdcbc7",
+    "#ffe6e6",
+  ],
+} as const;
