+++ packages/core/tests/crameri-ramps-diverging-data.ts
@@ -1,0 +1,137 @@
+/**
+ * Crameri diverging ramp fixtures — 11-stop samples
+ * from Scientific Colour Maps v8.0.1 (Zenodo DOI 10.5281/zenodo.1243862).
+ * Data only: never imports bun:test.
+ */
+export const DIVERGING_RAMPS: Readonly<Record<string, readonly string[]>> = {
+  bam: [
+    "#65024b",
+    "#9b3881",
+    "#c164a9",
+    "#dd9acb",
+    "#f0d1e8",
+    "#f6f1f0",
+    "#e2edd1",
+    "#adcd89",
+    "#709e4b",
+    "#407626",
+    "#0d4c00",
+  ],
+  berlin: [
+    "#9eb0ff",
+    "#60a5df",
+    "#3280a6",
+    "#20526a",
+    "#112732",
+    "#190c09",
+    "#371000",
+    "#5f1f0a",
+    "#964a36",
+    "#ca7b71",
+    "#ffadad",
+  ],
+  broc: [
+    "#2c1a4c",
+    "#284174",
+    "#3f6b99",
+    "#7697b7",
+    "#b3c5d7",
+    "#ebeeec",
+    "#dcdbb8",
+    "#b6b67c",
+    "#81814c",
+    "#505023",
+    "#262600",
+  ],
+  cork: [
+    "#2c194c",
+    "#284275",
+    "#3d6b98",
+    "#6f92b3",
+    "#adc1d4",
+    "#e6edec",
+    "#b7cfb7",
+    "#7ba77a",
+    "#438142",
+    "#195615",
+    "#0f2903",
+  ],
+  lisbon: [
+    "#e6e5ff",
+    "#a0b3d6",
+    "#6083ae",
+    "#2a537d",
+    "#132a42",
+    "#171919",
+    "#383522",
+    "#66603d",
+    "#9a9160",
+    "#cfc998",
+    "#ffffd9",
+  ],
+  managua: [
+    "#ffcf67",
+    "#df9d56",
+    "#c17449",
+    "#a0513e",
+    "#773339",
+    "#572949",
+    "#4c3d73",
+    "#5161a1",
+    "#5f89c3",
+    "#6fb6e2",
+    "#81e7ff",
+  ],
+  roma: [
+    "#7e1700",
+    "#984e14",
+    "#ac7726",
+    "#c1a343",
+    "#d2d484",
+    "#c0eac3",
+    "#89dad7",
+    "#4db3cf",
+    "#2d88be",
+    "#1e5fac",
+    "#033198",
+  ],
+  tofino: [
+    "#ded9ff",
+    "#98a8e1",
+    "#5777ba",
+    "#304a7b",
+    "#19253d",
+    "#0d1613",
+    "#18321a",
+    "#2b5b2f",
+    "#4a8d4b",
+    "#91be74",
+    "#dbe69b",
+  ],
+  vanimo: [
+    "#ffcdfd",
+    "#d786c8",
+    "#ad539a",
+    "#742e64",
+    "#33172c",
+    "#1a1513",
+    "#232c14",
+    "#40581f",
+    "#62872f",
+    "#8abc53",
+    "#befda5",
+  ],
+  vik: [
+    "#001261",
+    "#023a7b",
+    "#116496",
+    "#5496b7",
+    "#a7c9da",
+    "#ece5e0",
+    "#e1b8a0",
+    "#cd8961",
+    "#b75a26",
+    "#852206",
+    "#590008",
+  ],
+} as const;
