+++ packages/core/tests/crameri-ramps.test.ts
@@ -1,6 +1,7 @@
 /**
  * Crameri continuous ramp resolution — fixtures pin 11-stop samples
  * from Scientific Colour Maps v8.0.1 (Zenodo DOI 10.5281/zenodo.1243862).
+ * Ramp data lives in sibling crameri-ramps-*-data.ts modules, split by family.
  */
 import { describe, expect, it } from "bun:test";
 
@@ -12,471 +13,22 @@ import {
 import { crameriRampStops } from "../src/scales/crameri-ramps.js";
 import { sampleSequentialPalette, trainColor } from "../src/scales/train-color.js";
 import { VIRIDIS_RAMP_10 } from "../src/scales/viridis-ramp.js";
+import { CONTINUOUS_RAMPS } from "./crameri-ramps-continuous-data.js";
+import { DIVERGING_RAMPS } from "./crameri-ramps-diverging-data.js";
+import { MULTISEQUENTIAL_RAMPS } from "./crameri-ramps-multisequential-data.js";
 
-const RAMPS: Readonly<Record<string, readonly string[]>> = {
-  acton: [
-    "#260d40",
-    "#3c2c5c",
-    "#4f4775",
-    "#665e8a",
-    "#85648e",
-    "#a86690",
-    "#ca7199",
-    "#d890b3",
-    "#e2b2cf",
-    "#e9d1e7",
-    "#f0eafa",
-  ],
-  bamako: [
-    "#003b47",
-    "#0d4340",
-    "#1b4c37",
-    "#2e582b",
-    "#47681c",
-    "#637a0a",
-    "#818800",
-    "#a2930d",
-    "#c6ae39",
-    "#e5cb75",
-    "#ffe5ad",
-  ],
-  batlow: [
-    "#011959",
-    "#103d5f",
-    "#185562",
-    "#30685c",
-    "#577647",
-    "#828231",
-    "#b38e2f",
-    "#e09651",
-    "#fba689",
-    "#fdb9c2",
-    "#faccfa",
-  ],
-  batlowW: [
-    "#011959",
-    "#103d5f",
-    "#175462",
-    "#2d685e",
-    "#53784c",
-    "#7e8737",
-    "#b1993a",
-    "#daa66b",
-    "#f3b49e",
-    "#ffdddb",
-    "#fffefe",
-  ],
-  batlowK: [
-    "#04050a",
-    "#18293d",
-    "#304d5d",
-    "#47605c",
-    "#61704d",
-    "#86833c",
-    "#b7953c",
-    "#e4a15f",
-    "#f9ac92",
-    "#fdbbc5",
-    "#faccfa",
-  ],
-  bilbao: [
-    "#4c0001",
-    "#712227",
-    "#8e3f46",
-    "#9e5a55",
-    "#a46f5a",
-    "#a9825e",
-    "#ae9663",
-    "#b9ae80",
-    "#c4c0ac",
-    "#d6d5d3",
-    "#ffffff",
-  ],
-  buda: [
-    "#b301b3",
-    "#b3289f",
-    "#b84294",
-    "#bf598c",
-    "#c76f85",
-    "#cd857e",
-    "#d39b79",
-    "#d8b173",
-    "#dec96d",
-    "#e7e268",
-    "#ffff66",
-  ],
-  davos: [
-    "#00054a",
-    "#102a6f",
-    "#234a8c",
-    "#3a679b",
-    "#547d9c",
-    "#6c8e93",
-    "#849e89",
-    "#a4b68a",
-    "#d4dba8",
-    "#f5f5d7",
-    "#fefefe",
-  ],
-  devon: [
-    "#2c1a4c",
-    "#293265",
-    "#274a7e",
-    "#2f62a0",
-    "#4c77c4",
-    "#7e8fdd",
-    "#a8a5ec",
-    "#c0baf2",
-    "#d5d1f6",
-    "#eae8fb",
-    "#ffffff",
-  ],
-  glasgow: [
-    "#361338",
-    "#4d1924",
-    "#652211",
-    "#743b01",
-    "#735704",
-    "#6d702d",
-    "#65845d",
-    "#60988c",
-    "#7eaeb9",
-    "#abc0dc",
-    "#dbd3ff",
-  ],
-  grayC: [
-    "#000000",
-    "#212121",
-    "#383838",
-    "#4e4e4e",
-    "#636363",
-    "#777777",
-    "#8b8b8b",
-    "#a2a2a2",
-    "#bdbdbd",
-    "#dcdcdc",
-    "#ffffff",
-  ],
-  hawaii: [
-    "#8c0273",
-    "#91285a",
-    "#954147",
-    "#985936",
-    "#9c7524",
-    "#9c961c",
-    "#8fb63c",
-    "#78cb70",
-    "#62dca9",
-    "#71ece0",
-    "#b3f2fd",
-  ],
-  imola: [
-    "#1a33b3",
-    "#2345aa",
-    "#2c55a1",
-    "#356598",
-    "#42748b",
-    "#54867f",
-    "#6a9d78",
-    "#83b672",
-    "#9fd26b",
-    "#caec67",
-    "#ffff66",
-  ],
-  lajolla: [
-    "#191900",
-    "#31220e",
-    "#512d1e",
-    "#7d3b35",
-    "#b34947",
-    "#d9604e",
-    "#e38050",
-    "#e99d53",
-    "#f0bd57",
-    "#f9e384",
-    "#fffecb",
-  ],
-  lapaz: [
-    "#1a0c64",
-    "#222b79",
-    "#28468b",
-    "#315e99",
-    "#4277a2",
-    "#5c8ca3",
-    "#7c9b9e",
-    "#9fa696",
-    "#cab79d",
-    "#f2d7c6",
-    "#fef2f3",
-  ],
-  lipari: [
-    "#031326",
-    "#103557",
-    "#3c5478",
-    "#615e78",
-    "#806070",
-    "#a56267",
-    "#cf695e",
-    "#e98768",
-    "#e6aa7f",
-    "#eacda5",
-    "#fdf5da",
-  ],
-  navia: [
-    "#031327",
-    "#06315a",
-    "#105185",
-    "#236c91",
-    "#327c89",
-    "#418a80",
-    "#549b75",
-    "#70b369",
-    "#a7d278",
-    "#dde7ae",
-    "#fcf4d9",
-  ],
-  naviaW: [
-    "#041427",
-    "#08345c",
-    "#165584",
-    "#29708f",
-    "#388188",
-    "#4a927f",
-    "#62a774",
-    "#8cc574",
-    "#cbe6a3",
-    "#eff6db",
-    "#fefefd",
-  ],
-  nuuk: [
-    "#05598c",
-    "#276184",
-    "#436e82",
-    "#638089",
-    "#859493",
-    "#a1a698",
-    "#b2b293",
-    "#bdbd8a",
-    "#caca83",
-    "#e4e391",
-    "#fefeb2",
-  ],
-  oslo: [
-    "#010101",
-    "#0d1927",
-    "#122d48",
-    "#1a446d",
-    "#2c5d96",
-    "#507bbc",
-    "#7494c9",
-    "#92a6c9",
-    "#b2bccc",
-    "#d8dade",
-    "#ffffff",
-  ],
-  tokyo: [
-    "#1c0e34",
-    "#481e43",
-    "#653a4d",
-    "#6e5151",
-    "#716152",
-    "#747053",
-    "#788555",
-    "#7fa35c",
-    "#95cb78",
-    "#c7f0b2",
-    "#effcdd",
-  ],
-  turku: [
-    "#000000",
-    "#22221f",
-    "#3c3c32",
-    "#565640",
-    "#71704c",
-    "#938c5b",
-    "#b9a06e",
-    "#d8a782",
-    "#efb1a1",
-    "#fdcbc7",
-    "#ffe6e6",
-  ],
-  bam: [
-    "#65024b",
-    "#9b3881",
-    "#c164a9",
-    "#dd9acb",
-    "#f0d1e8",
-    "#f6f1f0",
-    "#e2edd1",
-    "#adcd89",
-    "#709e4b",
-    "#407626",
-    "#0d4c00",
-  ],
-  berlin: [
-    "#9eb0ff",
-    "#60a5df",
-    "#3280a6",
-    "#20526a",
-    "#112732",
-    "#190c09",
-    "#371000",
-    "#5f1f0a",
-    "#964a36",
-    "#ca7b71",
-    "#ffadad",
-  ],
-  broc: [
-    "#2c1a4c",
-    "#284174",
-    "#3f6b99",
-    "#7697b7",
-    "#b3c5d7",
-    "#ebeeec",
-    "#dcdbb8",
-    "#b6b67c",
-    "#81814c",
-    "#505023",
-    "#262600",
-  ],
-  cork: [
-    "#2c194c",
-    "#284275",
-    "#3d6b98",
-    "#6f92b3",
-    "#adc1d4",
-    "#e6edec",
-    "#b7cfb7",
-    "#7ba77a",
-    "#438142",
-    "#195615",
-    "#0f2903",
-  ],
-  lisbon: [
-    "#e6e5ff",
-    "#a0b3d6",
-    "#6083ae",
-    "#2a537d",
-    "#132a42",
-    "#171919",
-    "#383522",
-    "#66603d",
-    "#9a9160",
-    "#cfc998",
-    "#ffffd9",
-  ],
-  managua: [
-    "#ffcf67",
-    "#df9d56",
-    "#c17449",
-    "#a0513e",
-    "#773339",
-    "#572949",
-    "#4c3d73",
-    "#5161a1",
-    "#5f89c3",
-    "#6fb6e2",
-    "#81e7ff",
-  ],
-  roma: [
-    "#7e1700",
-    "#984e14",
-    "#ac7726",
-    "#c1a343",
-    "#d2d484",
-    "#c0eac3",
-    "#89dad7",
-    "#4db3cf",
-    "#2d88be",
-    "#1e5fac",
-    "#033198",
-  ],
-  tofino: [
-    "#ded9ff",
-    "#98a8e1",
-    "#5777ba",
-    "#304a7b",
-    "#19253d",
-    "#0d1613",
-    "#18321a",
-    "#2b5b2f",
-    "#4a8d4b",
-    "#91be74",
-    "#dbe69b",
-  ],
-  vanimo: [
-    "#ffcdfd",
-    "#d786c8",
-    "#ad539a",
-    "#742e64",
-    "#33172c",
-    "#1a1513",
-    "#232c14",
-    "#40581f",
-    "#62872f",
-    "#8abc53",
-    "#befda5",
-  ],
-  vik: [
-    "#001261",
-    "#023a7b",
-    "#116496",
-    "#5496b7",
-    "#a7c9da",
-    "#ece5e0",
-    "#e1b8a0",
-    "#cd8961",
-    "#b75a26",
-    "#852206",
-    "#590008",
-  ],
-  bukavu: [
-    "#1a3333",
-    "#214e71",
-    "#2f7ab6",
-    "#56a6c9",
-    "#8ad1cf",
-    "#014026",
-    "#326513",
-    "#7d7732",
-    "#b68f60",
-    "#d6c8b4",
-    "#ededfc",
-  ],
-  fes: [
-    "#0d0d0d",
-    "#3c3c3c",
-    "#646464",
-    "#888888",
-    "#b7b7b7",
-    "#024026",
-    "#4d5923",
-    "#88682e",
-    "#be905f",
-    "#d6c8b4",
-    "#ededfc",
-  ],
-  oleron: [
-    "#1a2659",
-    "#424f82",
-    "#6c79ac",
-    "#99a6d9",
-    "#c4d1f6",
-    "#1a4c00",
-    "#535e02",
-    "#8c7c31",
-    "#c5a46c",
-    "#f3d4ab",
-    "#fdfde6",
-  ],
-} as const;
+const ALL_RAMPS: Readonly<Record<string, readonly string[]>> = {
+  ...CONTINUOUS_RAMPS,
+  ...DIVERGING_RAMPS,
+  ...MULTISEQUENTIAL_RAMPS,
+};
 
 describe("Crameri Scientific colour maps", () => {
   it("registers the full continuous suite (35 maps)", () => {
-    expect(Object.keys(RAMPS).length).toBe(35);
+    expect(Object.keys(ALL_RAMPS).length).toBe(35);
   });
 
-  for (const [name, stops] of Object.entries(RAMPS)) {
+  for (const [name, stops] of Object.entries(CONTINUOUS_RAMPS)) {
     it(`resolves ${name} for continuous/binned scales`, () => {
       expect(crameriRampStops(name)).toEqual([...stops]);
       expect(continuousSchemeRamp(name)).toEqual([...stops]);
@@ -485,13 +37,13 @@ describe("Crameri Scientific colour maps", () => {
   }
 
   it("resolves a Crameri ramp for ordinal fingerprint fallthrough", () => {
-    expect(resolveOrdinalPaletteStops({ scheme: "batlow" })).toEqual([...RAMPS.batlow!]);
-    expect(resolveOrdinalPaletteStops({ scheme: "vik" })).toEqual([...RAMPS.vik!]);
+    expect(resolveOrdinalPaletteStops({ scheme: "batlow" })).toEqual([...CONTINUOUS_RAMPS.batlow!]);
+    expect(resolveOrdinalPaletteStops({ scheme: "vik" })).toEqual([...DIVERGING_RAMPS.vik!]);
   });
 
   it("even-samples Crameri ramps for discrete ordinal training (viridis_d parity)", () => {
     const scale = trainColor(["a", "b", "c", "d"], null, { scheme: "batlow" });
-    const expected = sampleSequentialPalette(RAMPS.batlow!, 4);
+    const expected = sampleSequentialPalette(CONTINUOUS_RAMPS.batlow!, 4);
     expect(scale.colorOf("a")).toBe(expected[0]);
     expect(scale.colorOf("d")).toBe(expected[3]);
   });
@@ -507,24 +59,5 @@ describe("Crameri Scientific colour maps", () => {
       VIRIDIS_RAMP_10,
     );
     expect(crameriRampStops("batlowS")).toBeUndefined();
-  });
-
-  it("resolves cyclic *O ramps as 11-stop sequential tables", () => {
-    const romaO = [
-      "#733957",
-      "#863f38",
-      "#9c5d2b",
-      "#b88e3b",
-      "#d3c876",
-      "#cbe1b3",
-      "#9bd4cd",
-      "#65adca",
-      "#4e7cb2",
-      "#5e4f85",
-      "#723959",
-    ];
-    expect(crameriRampStops("romaO")).toEqual(romaO);
-    expect(resolveSequentialPipelineRange({ scheme: "romaO" }, VIRIDIS_RAMP_10)).toEqual(romaO);
-    expect(continuousSchemeRamp("romaO")).toBeUndefined();
   });
 });
