+++ packages/core/tests/semantic-viewport.test.ts
@@ -3,7 +3,6 @@ import { describe, expect, it } from "bun:test";
 import { aes, gg } from "@ggsvelte/spec";
 
 import { runPipeline } from "../src/pipeline.ts";
-import { encodeKey } from "../src/scales/state.ts";
 
 describe("RenderModel semantic viewport", () => {
   it("inverts a plot-pixel rectangle through the panel's semantic axes", () => {
@@ -157,201 +156,6 @@ describe("RenderModel semantic viewport", () => {
     expect(pixels.y1).toBeCloseTo(scenePanel.y + scenePanel.height, 10);
   });
 
-  it("resolves a facet by pixel position and uses its panel-local scale", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { panel: "west", x: 0, y: 0 },
-          { panel: "west", x: 10, y: 1 },
-          { panel: "east", x: 100, y: 0 },
-          { panel: "east", x: 200, y: 1 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .facet({ wrap: "panel", scales: "free_x" })
-        .scales({ x: { nice: false, expand: { mult: 0, add: 0 } } })
-        .spec(),
-      { width: 600, height: 300 },
-    );
-    const scenePanel = model.scene.panels.find((panel) => panel.strip === "east")!;
-    const panel = model.viewport.panelAt({
-      x: scenePanel.x + scenePanel.width / 2,
-      y: scenePanel.y + scenePanel.height / 2,
-    })!;
-    const domains = panel.invert(panel.bounds);
-
-    expect(panel.id).toBe(scenePanel.id);
-    expect(model.viewport.panels.map((candidate) => candidate.id)).toEqual(
-      model.scene.panels.map((candidate) => candidate.id),
-    );
-    expect(domains.x?.[0]).toBeCloseTo(100, 10);
-    expect(domains.x?.[1]).toBeCloseTo(200, 10);
-  });
-
-  it("queries panel candidates in semantic-axis mode", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { x: 2, y: 1 },
-          { x: 8, y: 9 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .scales({
-          x: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 } },
-          y: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 } },
-        })
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-    const first = model.candidates.candidate(0)!;
-    const matches = panel.query(
-      {
-        x0: first.x - 1,
-        y0: scenePanel.y + scenePanel.height - 1,
-        x1: first.x + 1,
-        y1: scenePanel.y + scenePanel.height,
-      },
-      "x",
-    );
-
-    expect(matches.map((candidate) => candidate.rowIndex)).toEqual([0]);
-  });
-
-  it("projects encoded categorical identities edge to edge", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { category: "a", y: 1 },
-          { category: "b", y: 2 },
-          { category: "c", y: 3 },
-          { category: "d", y: 4 },
-        ],
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-    const pixels = panel.project({
-      x: { kind: "band", keys: [encodeKey("b"), encodeKey("c")] },
-    });
-
-    expect(pixels.x0).toBeCloseTo(scenePanel.x + scenePanel.width * 0.25, 10);
-    expect(pixels.x1).toBeCloseTo(scenePanel.x + scenePanel.width * 0.75, 10);
-  });
-
-  /**
-   * The band extent must not be derived via Math.min(...centers): `keys` is a
-   * public unbounded string[], so spreading one argument per selected key
-   * RangeErrors on a wide enough band brush. Every key here is distinct, so a
-   * dedupe-then-spread "fix" fails this too. Mirrors the grouping.ts guard.
-   */
-  it("projects a wide band selection without spreading the keys", () => {
-    const categories = Array.from({ length: 500 }, (_, i) => `c${String(i)}`);
-    const model = runPipeline(
-      gg(
-        categories.map((category, i) => ({ category, y: i })),
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-    const originalMin = Math.min;
-    const originalMax = Math.max;
-    const guard =
-      (name: string, original: (...args: number[]) => number) =>
-      (...args: number[]): number => {
-        if (args.length > 100) {
-          throw new Error(
-            `Math.${name} spread over ${String(args.length)} args (band extent leak)`,
-          );
-        }
-        return original(...args);
-      };
-    Math.min = guard("min", originalMin);
-    Math.max = guard("max", originalMax);
-    try {
-      const pixels = panel.project({
-        x: { kind: "band", keys: categories.map((category) => encodeKey(category)) },
-      });
-
-      expect(pixels.x0).toBeCloseTo(scenePanel.x, 10);
-      expect(pixels.x1).toBeCloseTo(scenePanel.x + scenePanel.width, 10);
-    } finally {
-      Math.min = originalMin;
-      Math.max = originalMax;
-    }
-  });
-
-  it("skips band keys outside the domain when projecting", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { category: "a", y: 1 },
-          { category: "b", y: 2 },
-          { category: "c", y: 3 },
-          { category: "d", y: 4 },
-        ],
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-
-    const mixed = panel.project({
-      x: { kind: "band", keys: [encodeKey("zz"), encodeKey("b"), encodeKey("c"), encodeKey("yy")] },
-    });
-    expect(mixed.x0).toBeCloseTo(scenePanel.x + scenePanel.width * 0.25, 10);
-    expect(mixed.x1).toBeCloseTo(scenePanel.x + scenePanel.width * 0.75, 10);
-
-    const unknown = panel.project({
-      x: { kind: "band", keys: [encodeKey("zz")] },
-    });
-    expect(unknown.x0).toBeCloseTo(scenePanel.x, 10);
-    expect(unknown.x1).toBeCloseTo(scenePanel.x + scenePanel.width, 10);
-
-    const empty = panel.project({ x: { kind: "band", keys: [] } });
-    expect(empty.x0).toBeCloseTo(scenePanel.x, 10);
-    expect(empty.x1).toBeCloseTo(scenePanel.x + scenePanel.width, 10);
-  });
-
-  it("resolves encoded categorical identities to raw semantic endpoints", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { category: 1, y: 1 },
-          { category: "1", y: 2 },
-          { category: true, y: 3 },
-        ],
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-
-    expect(
-      panel.resolve({
-        x: { kind: "band", keys: [encodeKey(1), encodeKey(true)] },
-      }).x,
-    ).toEqual([1, true]);
-  });
-
   it("reports raw screen-normalized spans for a plot-pixel rect", () => {
     const model = runPipeline(
       gg(
@@ -380,485 +184,5 @@ describe("RenderModel semantic viewport", () => {
 
     expect(span.x).toBeCloseTo(0.5, 10);
     expect(span.y).toBeCloseTo(0.5, 10);
-  });
-
-  it("exposes a continuous axis edit model with domain and reversal", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { x: 0, y: 0 },
-          { x: 10, y: 10 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .scales({
-          x: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 } },
-          y: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 }, reverse: true },
-        })
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-    const x = panel.axisEditModel("x");
-    const y = panel.axisEditModel("y");
-
-    expect(x).toEqual({
-      kind: "continuous",
-      type: "linear",
-      transform: "identity",
-      domain: [0, 10],
-      reversed: false,
-    });
-    expect(y.kind).toBe("continuous");
-    if (y.kind === "continuous") {
-      expect(y.type).toBe("linear");
-      expect(y.transform).toBe("identity");
-      expect(y.domain[0]).toBeCloseTo(0, 10);
-      expect(y.domain[1]).toBeCloseTo(10, 10);
-      expect(y.reversed).toBe(true);
-    }
-  });
-
-  it("slices inclusive band categories between edit endpoints", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { category: "a", y: 1 },
-          { category: "b", y: 2 },
-          { category: "c", y: 3 },
-          { category: "d", y: 4 },
-        ],
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-    const x = panel.axisEditModel("x");
-
-    expect(x.kind).toBe("band");
-    if (x.kind === "band") {
-      expect(x.rawDomain).toEqual(["a", "b", "c", "d"]);
-      expect(x.slice(["b", "c"])).toEqual(["b", "c"]);
-      expect(x.slice(["c", "a"])).toEqual(["a", "b", "c"]);
-      expect(x.slice(["b", "missing"])).toBeUndefined();
-    }
-  });
-
-  it("panel.nearest never surfaces a candidate from another facet (#787)", () => {
-    // Stacked facets + mode "x": unscoped nearest uses a full-height strip and
-    // |Δx| distance, so a hover in A at B's screen-x seeds B. Panel-scoped
-    // nearest must refuse that leak.
-    const model = runPipeline(
-      gg(
-        [
-          { g: "A", x: 1, y: 5 },
-          { g: "B", x: 5, y: 5 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .facet({ wrap: "g", ncol: 1 })
-        .scales({
-          x: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 } },
-          y: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 } },
-        })
-        .spec(),
-      { width: 200, height: 400 },
-    );
-    const panelA = model.scene.panels[0]!;
-    const candB = model.candidates.candidate(1)!;
-    expect(candB.panelId).toBe(model.scene.panels[1]!.id);
-
-    const probe = { x: candB.x, y: panelA.y + panelA.height / 2 };
-    // Documents the leak the viewport API closes.
-    expect(
-      model.candidates.nearest(probe.x, probe.y, { mode: "x", maxDistance: 24 })?.panelId,
-    ).toBe(candB.panelId);
-
-    const viewportPanel = model.viewport.panelAt(probe)!;
-    expect(viewportPanel.id).toBe(panelA.id);
-    expect(viewportPanel.nearest(probe, { mode: "x", maxDistance: 24 })).toBeNull();
-  });
-
-  it("panelAtOrOnly falls back only when the plot has a single panel", () => {
-    const single = runPipeline(
-      gg(
-        [
-          { x: 0, y: 0 },
-          { x: 10, y: 10 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .scales({
-          x: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 } },
-          y: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 } },
-        })
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const only = single.viewport.panels[0]!;
-    const outside = { x: only.bounds.x0 - 10, y: only.bounds.y0 - 10 };
-    expect(single.viewport.panelAt(outside)).toBeNull();
-    expect(single.viewport.panelAtOrOnly(outside)?.id).toBe(only.id);
-    expect(
-      single.viewport.panelAtOrOnly({ x: only.bounds.x0 + 1, y: only.bounds.y0 + 1 })?.id,
-    ).toBe(only.id);
-
-    const multi = runPipeline(
-      gg(
-        [
-          { g: "A", x: 1, y: 1 },
-          { g: "B", x: 2, y: 2 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .facet({ wrap: "g" })
-        .spec(),
-      { width: 400, height: 200 },
-    );
-    expect(multi.viewport.panels.length).toBe(2);
-    expect(multi.viewport.panelAtOrOnly({ x: -50, y: -50 })).toBeNull();
-    const first = multi.viewport.panels[0]!;
-    expect(
-      multi.viewport.panelAtOrOnly({
-        x: (first.bounds.x0 + first.bounds.x1) / 2,
-        y: (first.bounds.y0 + first.bounds.y1) / 2,
-      })?.id,
-    ).toBe(first.id);
-  });
-
-  it("locate maps client coordinates into scene space under CSS scale", () => {
-    // Scene 400×200; CSS-scaled capture rect 200×100 at (100,100) → 2× scale on both axes.
-    const model = runPipeline(
-      gg(
-        [
-          { x: 0, y: 0 },
-          { x: 10, y: 10 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .scales({
-          x: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 } },
-          y: { domain: [0, 10], nice: false, expand: { mult: 0, add: 0 } },
-        })
-        .spec(),
-      { width: 400, height: 200 },
-    );
-    expect(model.scene.width).toBe(400);
-    expect(model.scene.height).toBe(200);
-    expect(
-      model.viewport.locate(150, 120, { left: 100, top: 100, width: 200, height: 100 }),
-    ).toEqual({ x: 100, y: 40 });
-  });
-
-  it("locate accounts for non-zero rect offset and non-square CSS scale", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { x: 0, y: 0 },
-          { x: 10, y: 10 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 300, height: 150 },
-    );
-    // rect 100×50 → scale 3× on x, 3× on y; client at rect origin → plot origin.
-    expect(model.viewport.locate(40, 60, { left: 40, top: 60, width: 100, height: 50 })).toEqual({
-      x: 0,
-      y: 0,
-    });
-    // Mid-rect client: ((90-40)/100)*300 = 150, ((85-60)/50)*150 = 75.
-    expect(model.viewport.locate(90, 85, { left: 40, top: 60, width: 100, height: 50 })).toEqual({
-      x: 150,
-      y: 75,
-    });
-  });
-
-  it("locate returns origin when the capture element has zero size", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { x: 0, y: 0 },
-          { x: 10, y: 10 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 200 },
-    );
-    expect(model.viewport.locate(10, 10, { left: 0, top: 0, width: 0, height: 50 })).toEqual({
-      x: 0,
-      y: 0,
-    });
-    expect(model.viewport.locate(10, 10, { left: 0, top: 0, width: 50, height: 0 })).toEqual({
-      x: 0,
-      y: 0,
-    });
-  });
-
-  it("locate does not clamp out-of-bounds client coordinates", () => {
-    // Drag past the capture edge must keep negative / past-edge plot coords.
-    const model = runPipeline(
-      gg(
-        [
-          { x: 0, y: 0 },
-          { x: 10, y: 10 },
-        ],
-        aes({ x: "x", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 100, height: 100 },
-    );
-    expect(model.viewport.locate(50, 50, { left: 100, top: 100, width: 100, height: 100 })).toEqual(
-      {
-        x: -50,
-        y: -50,
-      },
-    );
-    expect(
-      model.viewport.locate(250, 250, { left: 100, top: 100, width: 100, height: 100 }),
-    ).toEqual({ x: 150, y: 150 });
-  });
-
-  it("resolves a band selection without materializing every key", () => {
-    // resolve() reports the first and last selected values. It used to map the
-    // whole key list to get them, allocating an array per key to read two.
-    const categories = Array.from({ length: 400 }, (_, i) => `c${i}`);
-    const model = runPipeline(
-      gg(
-        categories.map((category, i) => ({ category, y: i })),
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-
-    let reads = 0;
-    const keys = new Proxy(
-      categories.map((category) => encodeKey(category)),
-      {
-        get(target, prop) {
-          if (typeof prop === "string" && /^\d+$/.test(prop)) reads += 1;
-          return Reflect.get(target, prop) as unknown;
-        },
-      },
-    );
-    expect(panel.resolve({ x: { kind: "band", keys } }).x).toEqual(["c0", "c399"]);
-    // Exactly two, and independent of the key count: a bound relative to the
-    // fixture size would go green if the fixture ever shrank.
-    expect(reads).toBe(2);
-    model.dispose();
-  });
-
-  it("still finds the ends when the outermost keys are not on the axis", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { category: "a", y: 1 },
-          { category: "b", y: 2 },
-          { category: "c", y: 3 },
-        ],
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-
-    // Unknown keys bracket the real ones and must be skipped from both sides.
-    const keys = ["ghost-1", encodeKey("a"), encodeKey("c"), "ghost-2"];
-    expect(panel.resolve({ x: { kind: "band", keys } }).x).toEqual(["a", "c"]);
-    // Every key unknown: no span at all.
-    expect(panel.resolve({ x: { kind: "band", keys: ["ghost-1", "ghost-2"] } }).x).toBeUndefined();
-    // A single known key is both ends, whether or not unknowns bracket it —
-    // the second reaches the same return only after the backward scan runs.
-    expect(panel.resolve({ x: { kind: "band", keys: [encodeKey("b")] } }).x).toEqual(["b", "b"]);
-    expect(
-      panel.resolve({ x: { kind: "band", keys: ["ghost-1", encodeKey("b"), "ghost-2"] } }).x,
-    ).toEqual(["b", "b"]);
-    // No keys at all: nothing to span.
-    expect(panel.resolve({ x: { kind: "band", keys: [] } }).x).toBeUndefined();
-    // A key repeated is still just that value at both ends.
-    expect(
-      panel.resolve({ x: { kind: "band", keys: [encodeKey("b"), encodeKey("b")] } }).x,
-    ).toEqual(["b", "b"]);
-  });
-
-  it("spans a band y axis the same way", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { x: 1, category: "a" },
-          { x: 2, category: "b" },
-          { x: 3, category: "c" },
-        ],
-        aes({ x: "x", y: "category" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-    expect(
-      panel.resolve({ y: { kind: "band", keys: ["ghost", encodeKey("a"), encodeKey("c")] } }).y,
-    ).toEqual(["a", "c"]);
-    model.dispose();
-    model.dispose();
-  });
-
-  // --- #1332: project() normalizes only the domain extremes, not every key ---
-
-  it("projects a band selection by normalizing only the domain extremes (#1332)", () => {
-    const categories = Array.from({ length: 200 }, (_, i) => `c${i}`);
-    const model = runPipeline(
-      gg(
-        categories.map((category, i) => ({ category, y: i })),
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scale = model.scales.panels[0]?.x ?? model.scales.x;
-    let normalizeCalls = 0;
-    const original = scale.normalize.bind(scale);
-    scale.normalize = (value: unknown) => {
-      normalizeCalls += 1;
-      return original(value);
-    };
-    const panel = model.viewport.panel(model.scene.panels[0]!.id)!;
-    const keys = categories.map((category) => encodeKey(category));
-    const pixels = panel.project({ x: { kind: "band", keys } });
-    // Contiguous full domain: two extremes, one normalize each.
-    expect(normalizeCalls).toBe(2);
-    expect(pixels.x0).toBeCloseTo(model.scene.panels[0]!.x, 10);
-    expect(pixels.x1).toBeCloseTo(model.scene.panels[0]!.x + model.scene.panels[0]!.width, 10);
-    model.dispose();
-  });
-
-  it("projects non-contiguous band keys from domain extremes, not list ends (#1332)", () => {
-    // Keys ordered b, a, d — list ends are b and d, but the selection's domain
-    // extremes are a and d. The span must cover a→d.
-    const model = runPipeline(
-      gg(
-        [
-          { category: "a", y: 1 },
-          { category: "b", y: 2 },
-          { category: "c", y: 3 },
-          { category: "d", y: 4 },
-        ],
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-    const shuffled = panel.project({
-      x: {
-        kind: "band",
-        keys: [encodeKey("b"), encodeKey("a"), encodeKey("d")],
-      },
-    });
-    const full = panel.project({
-      x: {
-        kind: "band",
-        keys: [encodeKey("a"), encodeKey("b"), encodeKey("c"), encodeKey("d")],
-      },
-    });
-    // a is the leftmost category; d the rightmost. Missing c does not shrink
-    // the outer extent when a and d are both selected.
-    expect(shuffled.x0).toBeCloseTo(full.x0, 10);
-    expect(shuffled.x1).toBeCloseTo(full.x1, 10);
-    expect(shuffled.x0).toBeCloseTo(scenePanel.x, 10);
-    expect(shuffled.x1).toBeCloseTo(scenePanel.x + scenePanel.width, 10);
-    model.dispose();
-  });
-
-  it("projects a reversed band axis from the same extremes (#1332)", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { category: "a", y: 1 },
-          { category: "b", y: 2 },
-          { category: "c", y: 3 },
-          { category: "d", y: 4 },
-        ],
-        aes({ x: "category", y: "y" }),
-      )
-        .geomPoint()
-        .scales({ x: { reverse: true } })
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-    // Contiguous b,c under reverse: a is rightmost in screen space.
-    const pixels = panel.project({
-      x: { kind: "band", keys: [encodeKey("b"), encodeKey("c")] },
-    });
-    // Domain indices 1 and 2 of 4; reverse flips centers. Edge-to-edge of b,c
-    // is the middle half of the panel still — reverse swaps left/right but
-    // the expanded half-step span width is the same.
-    expect(pixels.x1 - pixels.x0).toBeCloseTo(scenePanel.width * 0.5, 10);
-    expect(pixels.x0).toBeCloseTo(scenePanel.x + scenePanel.width * 0.25, 10);
-    expect(pixels.x1).toBeCloseTo(scenePanel.x + scenePanel.width * 0.75, 10);
-    model.dispose();
-  });
-
-  it("projects a band y selection the same way (#1332)", () => {
-    const model = runPipeline(
-      gg(
-        [
-          { x: 1, category: "a" },
-          { x: 2, category: "b" },
-          { x: 3, category: "c" },
-          { x: 4, category: "d" },
-        ],
-        aes({ x: "x", y: "category" }),
-      )
-        .geomPoint()
-        .spec(),
-      { width: 400, height: 300 },
-    );
-    const scenePanel = model.scene.panels[0]!;
-    const panel = model.viewport.panel(scenePanel.id)!;
-    const scale = model.scales.panels[0]?.y ?? model.scales.y;
-    let normalizeCalls = 0;
-    const original = scale.normalize.bind(scale);
-    scale.normalize = (value: unknown) => {
-      normalizeCalls += 1;
-      return original(value);
-    };
-    const pixels = panel.project({
-      y: {
-        kind: "band",
-        keys: [encodeKey("d"), encodeKey("a"), encodeKey("c")],
-      },
-    });
-    expect(normalizeCalls).toBe(2);
-    // a..d cover the full y band domain.
-    expect(pixels.y0).toBeCloseTo(scenePanel.y, 10);
-    expect(pixels.y1).toBeCloseTo(scenePanel.y + scenePanel.height, 10);
-    model.dispose();
   });
 });
