+++ scripts/guards/max-loc-exclusions.txt
@@ -16,8 +16,6 @@ packages/svelte/src/lib/index.ts 1086 # legacy oversize; pending refactor slice 
 packages/spec/src/index.ts 1018 # legacy oversize; pending refactor slice S3
 examples/hex/basic/data.ts 1014 # pure data corpus, no logic — permanently exempt
 scripts/release-wiring.test.ts 1005 # legacy oversize; pending refactor slice S7
-packages/core/tests/pipeline-responsive-guides.test.ts 999 # legacy oversize; pending refactor slice S2
-packages/core/tests/pipeline-m2/geom-sf.test.ts 986 # legacy oversize; pending refactor slice S2
 examples/density/overlay/data.ts 948 # pure data corpus, no logic — permanently exempt
 examples/errorbar/summary-bin/data.ts 945 # pure data corpus, no logic — permanently exempt
 examples/point/count/data.ts 943 # pure data corpus, no logic — permanently exempt
@@ -25,7 +23,6 @@ packages/svelte/tests/inspection/resolve-snapshot.test.ts 940 # legacy oversize;
 packages/svelte/src/lib/data/kyoto-sakura.ts 891 # pure data corpus, no logic — permanently exempt
 packages/svelte/src/lib/plot-engine.svelte.ts 883 # legacy oversize; pending refactor slice S5
 scripts/sakura-lesson.test.ts 870 # legacy oversize; pending refactor slice S7
-packages/core/tests/semantic-viewport.test.ts 864 # legacy oversize; pending refactor slice S2
 packages/spec/tests/validate-tier2-color.test.ts 862 # legacy oversize; pending refactor slice S4
 apps/docs/src/lib/theme-specimens/data.ts 821 # pure data corpus, no logic — permanently exempt
 examples/point/stat-unique/data.ts 796 # pure data corpus, no logic — permanently exempt
@@ -42,7 +39,6 @@ packages/svelte/tests/assembly/assemble.test.ts 652 # legacy oversize; pending r
 scripts/ci-routing/routing.ts 638 # legacy oversize; pending refactor slice S7
 packages/svelte/tests/interval/consumption.test.ts 628 # legacy oversize; pending refactor slice S6
 packages/svelte/tests/legend/filter-state.test.ts 617 # legacy oversize; pending refactor slice S6
-packages/core/tests/pipeline-layer-data.test.ts 613 # legacy oversize; pending refactor slice S2
 examples/density/kde-2d/data.ts 611 # pure data corpus, no logic — permanently exempt
 packages/svelte/src/lib/inspection/inspection-state.svelte.ts 585 # legacy oversize; pending refactor slice S5
 packages/svelte/src/lib/interval/interval-state.svelte.ts 583 # legacy oversize; pending refactor slice S5
@@ -57,7 +53,6 @@ packages/svelte/tests/inspection/inspection-state.pin-queue.test.ts 544 # legacy
 packages/svelte/src/lib/legend/focus-state.svelte.ts 543 # legacy oversize; pending refactor slice S5
 packages/svelte/tests/runtime/semantic-candidate-projection.test.ts 543 # legacy oversize; pending refactor slice S6
 examples/point/gradient-continuous/data.ts 530 # pure data corpus, no logic — permanently exempt
-packages/core/tests/crameri-ramps.test.ts 530 # legacy oversize; pending refactor slice S2
 packages/svelte/src/lib/scene/Batch.svelte 526 # legacy oversize; pending refactor slice S5
 benchmarks/competitive/fixtures/main.ts 523 # legacy oversize; pending refactor slice S7
 packages/spec/src/capabilities.ts 522 # legacy oversize; pending refactor slice S4
